oops finish moving
